make suggested changes
