--- packages/jest-environment-miniflare/test/fixtures/mock-worker.js
@@ -1,5 +1,0 @@
-export default {
-  async fetch(request) {
-    return await fetch(request);
-  },
-};
